requirements.txt
pandas==1.5.3
numpy==1.23.5
scipy==1.10.1
scikit-learn==1.2.2
plotly==5.14.1
nbformat==5.7.3
ipywidgets==7.7.5
PyYAML==6.0
streamlit==1.24.1
